# specs/004-app-logic-docs/contracts/openapi-excerpt.yaml
openapi: 3.0.3
info:
  title: Stock Verify API (Excerpt)
  version: "v-current"
  description: |
    Minimal OpenAPI excerpt for documentation purposes.
    This file is not intended to be complete or authoritative; validate against `/api/docs`.
servers:
  - url: http://localhost:8001
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Session:
      type: object
      properties:
        id: { type: string }
        warehouse: { type: string }
        staff_user: { type: string }
        staff_name: { type: string }
        type: { type: string }
        status: { type: string }
        started_at: { type: string, format: date-time }
    PaginatedSessions:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Session' }
        total: { type: integer }
        page: { type: integer }
        page_size: { type: integer }
        total_pages: { type: integer }
        has_next: { type: boolean }
        has_previous: { type: boolean }
paths:
...
  /api/sessions:
    get:
      summary: List sessions (paginated)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: page_size
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        '200':
          description: Paginated sessions
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaginatedSessions' }
    post:
      summary: Create session
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                warehouse: { type: string }
                type: { type: string }
              required: [warehouse]
      responses:
        '200':
          description: Created session
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Session' }
...